requirements.txt
sentence-transformers==3.1.1
pinecone==5.3.1
apache-airflow==2.8.1
 